# archforge-server-admin/src/main/java/com/lesofn/archforge/server/admin/controller/blog/BlogArticleController.java
package com.lesofn.archforge.server.admin.controller.blog;

import com.lesofn.archforge.blog.api.domain.BlogArticle;
import org.jspecify.annotations.Nullable;
import com.lesofn.archforge.blog.api.domain.BlogCategory;
import com.lesofn.archforge.blog.api.enums.BlogArticleStatus;
import com.lesofn.archforge.blog.api.service.BlogArticleService;
import com.lesofn.archforge.blog.api.service.BlogCategoryService;
import com.lesofn.archforge.infrastructure.auth.LoginContext;
import com.lesofn.archforge.common.auth.SystemLoginUser;
import com.lesofn.archforge.server.admin.dto.AdminPageResponse;
import com.lesofn.archforge.server.admin.dto.request.AdminBlogArticleCreateRequest;
import com.lesofn.archforge.server.admin.dto.request.AdminBlogArticleListRequest;
import com.lesofn.archforge.server.admin.dto.request.AdminBlogArticleUpdateRequest;
import com.lesofn.archforge.server.admin.dto.request.IdRequest;
import com.lesofn.archforge.server.admin.dto.response.AdminBlogArticleResponse;
import com.lesofn.archforge.server.admin.convertor.AdminBlogConvertor;
import jakarta.validation.Valid;
import java.util.List;
import java.util.stream.Collectors;
import lombok.RequiredArgsConstructor;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Sort;
import cn.dev33.satoken.annotation.SaCheckLogin;
import cn.dev33.satoken.annotation.SaCheckPermission;
import cn.dev33.satoken.annotation.SaCheckRole;
import com.lesofn.archforge.infrastructure.auth.stp.StpAdminUtil;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/admin/blog/article")
@SaCheckLogin(type = StpAdminUtil.TYPE)
@SaCheckRole(value = "ADMIN", type = StpAdminUtil.TYPE)
@RequiredArgsConstructor
public class BlogArticleController {

    private final BlogArticleService articleService;
    private final BlogCategoryService categoryService;
    private final AdminBlogConvertor adminBlogMapper;

    @Value("${arch-forge.web.public-url:http://localhost:8081}")
    private String webPublicUrl;

    @PostMapping
    public AdminPageResponse<AdminBlogArticleResponse> list(@RequestBody @Valid AdminBlogArticleListRequest request) {
        PageRequest pageRequest = PageRequest.of(
                Math.max(request.getCurrentPage() - 1, 0),
                request.getPageSize(),
                Sort.by("createTime").descending());
        BlogArticleStatus status = toStatus(request.getStatus());
        Page<BlogArticle> page = articleService.pageAll(
                pageRequest, request.getCategoryId(), request.getTitle(), status);
        List<AdminBlogArticleResponse> list = page.getContent().stream()
                .map(this::toResponse)
                .collect(Collectors.toList());
        return AdminPageResponse.of(list, page.getTotalElements(), page.getSize(), request.getCurrentPage());
    }

    @SaCheckPermission(value = "blog:article:add", type = StpAdminUtil.TYPE)
    @PostMapping("/create")
    public Long create(@RequestBody @Valid AdminBlogArticleCreateRequest request) {
        SystemLoginUser loginUser = LoginContext.getAdminUser();
        BlogArticle article = buildFromRequest(request)
                .setAuthorId(loginUser.getUserId());
        return articleService.create(article).getId();
    }

    @SaCheckPermission(value = "blog:article:edit", type = StpAdminUtil.TYPE)
    @PutMapping("/update")
    public Boolean update(@RequestBody @Valid AdminBlogArticleUpdateRequest request) {
        BlogArticle article = buildFromRequest(request).setId(request.getId());
        articleService.update(article);
        return true;
    }

    @SaCheckPermission(value = "blog:article:remove", type = StpAdminUtil.TYPE)
    @PostMapping("/delete")
    public Boolean delete(@RequestBody @Valid IdRequest request) {
        articleService.delete(request.getId());
        return true;
    }

    @SaCheckPermission(value = "blog:article:publish", type = StpAdminUtil.TYPE)
    @PostMapping("/{id}/publish")
    public Boolean publish(@PathVariable Long id) {
        articleService.publish(id);
        return true;
    }

    @SaCheckPermission(value = "blog:article:offline", type = StpAdminUtil.TYPE)
    @PostMapping("/{id}/offline")
    public Boolean offline(@PathVariable Long id) {
        articleService.offline(id);
        return true;
    }

    private @Nullable BlogArticleStatus toStatus(@Nullable Integer value) {
        if (value == null) {
            return null;
        }
        return switch (value) {
            case 0 -> BlogArticleStatus.DRAFT;
            case 1 -> BlogArticleStatus.PUBLISHED;
            case 2 -> BlogArticleStatus.OFFLINE;
            default -> throw new IllegalArgumentException("状态值无效: " + value);
        };
    }

    private BlogArticle buildFromRequest(AdminBlogArticleCreateRequest request) {
        return new BlogArticle()
                .setCategoryId(request.getCategoryId())
                .setTitle(request.getTitle())
                .setSlug(request.getSlug())
                .setSummary(request.getSummary())
                .setContent(request.getContent())
                .setCoverImageFileId(request.getCoverImageFileId())
                .setStatus(java.util.Objects.requireNonNullElse(toStatus(request.getStatus()), BlogArticleStatus.DRAFT));
    }

    private BlogArticle buildFromRequest(AdminBlogArticleUpdateRequest request) {
        return new BlogArticle()
                .setCategoryId(request.getCategoryId())
                .setTitle(request.getTitle())
                .setSlug(request.getSlug())
                .setSummary(request.getSummary())
                .setContent(request.getContent())
                .setCoverImageFileId(request.getCoverImageFileId())
                .setStatus(java.util.Objects.requireNonNullElse(toStatus(request.getStatus()), BlogArticleStatus.DRAFT));
    }

    private AdminBlogArticleResponse toResponse(BlogArticle article) {
        AdminBlogArticleResponse response = adminBlogMapper.toArticleResponse(article);
        BlogCategory category = categoryService.findById(article.getCategoryId()).orElse(null);
        response.setCategoryName(category != null ? category.getName() : "");
        response.setStatusLabel(article.getStatus().getLabel());
        response.setCoverImageUrl(article.getCoverImageFileId() != null
                ? webPublicUrl + "/web/file/" + article.getCoverImageFileId()
                : "");
        return response;
    }
}