# archforge-server-admin/src/main/java/com/lesofn/archforge/server/admin/controller/blog/BlogCategoryController.java
package com.lesofn.archforge.server.admin.controller.blog;

import com.lesofn.archforge.blog.api.domain.BlogCategory;
import com.lesofn.archforge.blog.api.service.BlogArticleService;
import com.lesofn.archforge.blog.api.service.BlogCategoryService;
import com.lesofn.archforge.server.admin.dto.AdminPageResponse;
import com.lesofn.archforge.server.admin.dto.request.AdminBlogCategoryCreateRequest;
import com.lesofn.archforge.server.admin.dto.request.AdminBlogCategoryListRequest;
import com.lesofn.archforge.server.admin.dto.request.AdminBlogCategoryUpdateRequest;
import com.lesofn.archforge.server.admin.dto.request.IdRequest;
import com.lesofn.archforge.server.admin.dto.response.AdminBlogCategoryResponse;
import com.lesofn.archforge.server.admin.convertor.AdminBlogConvertor;
import jakarta.validation.Valid;
import java.util.List;
import java.util.stream.Collectors;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Sort;
import cn.dev33.satoken.annotation.SaCheckLogin;
import cn.dev33.satoken.annotation.SaCheckPermission;
import cn.dev33.satoken.annotation.SaCheckRole;
import com.lesofn.archforge.infrastructure.auth.stp.StpAdminUtil;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/admin/blog/category")
@SaCheckLogin(type = StpAdminUtil.TYPE)
@SaCheckRole(value = "ADMIN", type = StpAdminUtil.TYPE)
@RequiredArgsConstructor
public class BlogCategoryController {

    private final BlogCategoryService categoryService;
    private final BlogArticleService articleService;
    private final AdminBlogConvertor adminBlogMapper;

    @PostMapping
    public AdminPageResponse<AdminBlogCategoryResponse> list(@RequestBody @Valid AdminBlogCategoryListRequest request) {
        PageRequest pageRequest = PageRequest.of(
                Math.max(request.getCurrentPage() - 1, 0),
                request.getPageSize(),
                Sort.by("sortOrder").ascending());
        Page<BlogCategory> page = categoryService.page(pageRequest, request.getName());
        List<AdminBlogCategoryResponse> list = page.getContent().stream()
                .map(this::toResponse)
                .collect(Collectors.toList());
        return AdminPageResponse.of(list, page.getTotalElements(), page.getSize(), request.getCurrentPage());
    }

    @SaCheckPermission(value = "blog:category:add", type = StpAdminUtil.TYPE)
    @PostMapping("/create")
    public Long create(@RequestBody @Valid AdminBlogCategoryCreateRequest request) {
        BlogCategory category = new BlogCategory()
                .setName(request.getName())
                .setSlug(request.getSlug())
                .setSortOrder(java.util.Objects.requireNonNullElse(request.getSortOrder(), 0))
                .setStatus(java.util.Objects.requireNonNullElse(request.getStatus(), 1));
        return categoryService.create(category).getId();
    }

    @SaCheckPermission(value = "blog:category:edit", type = StpAdminUtil.TYPE)
    @PutMapping("/update")
    public Boolean update(@RequestBody @Valid AdminBlogCategoryUpdateRequest request) {
        BlogCategory category = new BlogCategory()
                .setId(request.getId())
                .setName(request.getName())
                .setSlug(request.getSlug())
                .setSortOrder(java.util.Objects.requireNonNullElse(request.getSortOrder(), 0))
                .setStatus(java.util.Objects.requireNonNullElse(request.getStatus(), 1));
        categoryService.update(category);
        return true;
    }

    @SaCheckPermission(value = "blog:category:remove", type = StpAdminUtil.TYPE)
    @PostMapping("/delete")
    public Boolean delete(@RequestBody @Valid IdRequest request) {
        categoryService.delete(request.getId());
        return true;
    }

    private AdminBlogCategoryResponse toResponse(BlogCategory category) {
        AdminBlogCategoryResponse response = adminBlogMapper.toCategoryResponse(category);
        long articleCount = articleService.pageAll(
                PageRequest.of(0, 1, Sort.by("createTime").descending()),
                category.getId(), null, null).getTotalElements();
        response.setStatusLabel(category.isVisible() ? "显示" : "隐藏");
        response.setArticleCount(articleCount);
        return response;
    }
}