# spec/openapi.yaml
openapi: "3.1.0"
info:
  title: "标题：ArchForge接口文档"
  description: "描述：文档说明"
  contact:
    name: "ArchForge"
  version: "版本号:1.0.0"
  license:
    name: "Apache-2.0"
    identifier: "Apache-2.0"
servers:
- url: "/"
security:
- bearer: []
...
paths:
...
  /admin/blog/article:
    post:
      tags:
      - "blog-article-controller"
      operationId: "list_2"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlogArticleListRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPageResponseAdminBlogArticleResponse"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "List 2"
  /admin/blog/article/create:
    post:
      tags:
      - "blog-article-controller"
      operationId: "create_2"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlogArticleCreateRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "integer"
                format: "int64"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Create 2"
  /admin/blog/article/delete:
    post:
      tags:
      - "blog-article-controller"
      operationId: "delete_2"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Delete 2"
  /admin/blog/article/update:
    put:
      tags:
      - "blog-article-controller"
      operationId: "update_3"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlogArticleUpdateRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Update 3"
  /admin/blog/article/{id}/offline:
    post:
      tags:
      - "blog-article-controller"
      operationId: "offline"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Offline"
  /admin/blog/article/{id}/publish:
    post:
      tags:
      - "blog-article-controller"
      operationId: "publish"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Publish"
  /admin/blog/category:
    post:
      tags:
      - "blog-category-controller"
      operationId: "list_1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlogCategoryListRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPageResponseAdminBlogCategoryResponse"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "List 1"
  /admin/blog/category/create:
    post:
      tags:
      - "blog-category-controller"
      operationId: "create_1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlogCategoryCreateRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "integer"
                format: "int64"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Create 1"
  /admin/blog/category/delete:
    post:
      tags:
      - "blog-category-controller"
      operationId: "delete_1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Delete 1"
  /admin/blog/category/update:
    put:
      tags:
      - "blog-category-controller"
      operationId: "update_2"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlogCategoryUpdateRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Update 2"
...
  /web/articles/{slug}:
    get:
      tags:
      - "web-blog-controller"
      operationId: "getArticle"
      parameters:
      - name: "slug"
        in: "path"
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebArticleDetailResponse"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Get article"
...
  /web/file/{fileId}:
    get:
      tags:
      - "web-file-controller"
      operationId: "download"
      parameters:
      - name: "fileId"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
                format: "binary"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Download"
...
components:
  schemas:
    AdminBlogArticleCreateRequest:
      required:
      - "categoryId"
      - "content"
      - "slug"
      - "status"
      - "title"
      type: "object"
      properties:
        categoryId:
          type: "integer"
          format: "int64"
        title:
          minLength: 1
          type: "string"
        slug:
          minLength: 1
          type: "string"
        summary:
          type: "string"
        content:
          minLength: 1
          type: "string"
        coverImageFileId:
          type: "integer"
          format: "int64"
        status:
          type: "integer"
          format: "int32"
    AdminBlogArticleListRequest:
      type: "object"
      properties:
        currentPage:
          type: "integer"
          format: "int32"
        pageSize:
          type: "integer"
          format: "int32"
        categoryId:
          type: "integer"
          format: "int64"
        title:
          type: "string"
        status:
          type: "integer"
          format: "int32"
    AdminBlogArticleResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        categoryId:
          type: "integer"
          format: "int64"
        categoryName:
          type: "string"
        title:
          type: "string"
        slug:
          type: "string"
        summary:
          type: "string"
        content:
          type: "string"
        coverImageFileId:
          type: "integer"
          format: "int64"
        coverImageUrl:
          type: "string"
        status:
          type: "integer"
          format: "int32"
        statusLabel:
          type: "string"
        publishTime:
          type: "string"
          format: "date-time"
        createTime:
          type: "string"
          format: "date-time"
        updateTime:
          type: "string"
          format: "date-time"
    AdminBlogArticleUpdateRequest:
      required:
      - "categoryId"
      - "content"
      - "id"
      - "slug"
      - "status"
      - "title"
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        categoryId:
          type: "integer"
          format: "int64"
        title:
          minLength: 1
          type: "string"
        slug:
          minLength: 1
          type: "string"
        summary:
          type: "string"
        content:
          minLength: 1
          type: "string"
        coverImageFileId:
          type: "integer"
          format: "int64"
        status:
          type: "integer"
          format: "int32"
    AdminBlogCategoryCreateRequest:
      required:
      - "name"
      - "slug"
      type: "object"
      properties:
        name:
          minLength: 1
          type: "string"
        slug:
          minLength: 1
          type: "string"
        sortOrder:
          type: "integer"
          format: "int32"
        status:
          type: "integer"
          format: "int32"
    AdminBlogCategoryListRequest:
      type: "object"
      properties:
        currentPage:
          type: "integer"
          format: "int32"
        pageSize:
          type: "integer"
          format: "int32"
        name:
          type: "string"
    AdminBlogCategoryResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        slug:
          type: "string"
        sortOrder:
          type: "integer"
          format: "int32"
        status:
          type: "integer"
          format: "int32"
        statusLabel:
          type: "string"
        articleCount:
          type: "integer"
          format: "int64"
        createTime:
          type: "string"
          format: "date-time"
    AdminBlogCategoryUpdateRequest:
      required:
      - "id"
      - "name"
      - "slug"
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          minLength: 1
          type: "string"
        slug:
          minLength: 1
          type: "string"
        sortOrder:
          type: "integer"
          format: "int32"
        status:
          type: "integer"
          format: "int32"
...
    AdminPageResponseAdminBlogArticleResponse:
      type: "object"
      properties:
        list:
          type: "array"
          items:
            $ref: "#/components/schemas/AdminBlogArticleResponse"
        total:
          type: "integer"
          format: "int64"
        pageSize:
          type: "integer"
          format: "int32"
        currentPage:
          type: "integer"
          format: "int32"
    AdminPageResponseAdminBlogCategoryResponse:
      type: "object"
      properties:
        list:
          type: "array"
          items:
            $ref: "#/components/schemas/AdminBlogCategoryResponse"
        total:
          type: "integer"
          format: "int64"
        pageSize:
          type: "integer"
          format: "int32"
        currentPage:
          type: "integer"
          format: "int32"
...
    IdRequest:
      required:
      - "id"
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
...
    WebArticleDetailResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        title:
          type: "string"
        slug:
          type: "string"
        summary:
          type: "string"
        content:
          type: "string"
        coverImageFileId:
          type: "integer"
          format: "int64"
        coverImageUrl:
          type: "string"
        categoryId:
          type: "integer"
          format: "int64"
        categoryName:
          type: "string"
        categorySlug:
          type: "string"
        publishTime:
          type: "string"
          format: "date-time"
        createTime:
          type: "string"
          format: "date-time"
...
  securitySchemes:
    bearer:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  responses:
    Error:
      description: "Error response (RFC 9457 ProblemDetail)"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              type:
                type: "string"
              title:
                type: "string"
              status:
                type: "integer"
              detail:
                type: "string"
              instance:
                type: "string"